One big step of the messy part of libmpg123 merge. The hard code mixes should be done, easy code mixes to follow... then the fun with the build system.
git-svn-id: svn://scm.orgis.org/mpg123/trunk@1000 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/xfermem.c
@@ -12,7 +12,7 @@
 
 #ifndef NOXFERMEM
 
-#include "mpg123.h"
+#include "mpg123app.h"
 #include <string.h>
 #include <errno.h>
 #include <sys/time.h>
@@ -25,7 +25,6 @@
 #include <sys/select.h>
 #endif
 
-
 #ifndef HAVE_MMAP
 #include <sys/ipc.h>
 #include <sys/shm.h>
@@ -35,9 +34,9 @@
 #define MAP_ANON MAP_ANONYMOUS
 #endif
 
-void xfermem_init (txfermem **xf, int bufsize, int msize, int skipbuf)
+void xfermem_init (txfermem **xf, size_t bufsize, size_t msize, size_t skipbuf)
 {
-	int regsize = bufsize + msize + skipbuf + sizeof(txfermem);
+	size_t regsize = bufsize + msize + skipbuf + sizeof(txfermem);
 
 #ifdef HAVE_MMAP
 #  ifdef MAP_ANON
@@ -116,9 +115,9 @@ void xfermem_init_reader (txfermem *xf)
 		close (xf->fd[XF_WRITER]);
 }
 
-int xfermem_get_freespace (txfermem *xf)
+size_t xfermem_get_freespace (txfermem *xf)
 {
-	int freeindex, readindex;
+	size_t freeindex, readindex;
 
 	if(!xf)
 		return 0;
@@ -132,9 +131,9 @@ int xfermem_get_freespace (txfermem *xf)
 		return ((xf->size - (freeindex - readindex)) - 1);
 }
 
-int xfermem_get_usedspace (txfermem *xf)
+size_t xfermem_get_usedspace (txfermem *xf)
 {
-	int freeindex, readindex;
+	size_t freeindex, readindex;
 
 	if(!xf)
 		return 0;
@@ -221,10 +220,10 @@ int xfermem_block (int readwrite, txfermem *xf)
 
 #else /* stubs for generic / win32 */
 
-#include "mpg123.h"
+#include "mpg123app.h"
 #include "xfermem.h"
 
-void xfermem_init (txfermem **xf, int bufsize, int msize, int skipbuf)
+void xfermem_init (txfermem **xf, size_t bufsize, size_t msize, size_t skipbuf)
 {
 }
 void xfermem_done (txfermem *xf)
@@ -236,11 +235,11 @@ void xfermem_init_writer (txfermem *xf)
 void xfermem_init_reader (txfermem *xf)
 {
 }
-int xfermem_get_freespace (txfermem *xf)
+size_t xfermem_get_freespace (txfermem *xf)
 {
   return 0;
 }
-int xfermem_get_usedspace (txfermem *xf)
+size_t xfermem_get_usedspace (txfermem *xf)
 {
   return 0;
 }
